+++ veilid-flutter/example/integration_test/test_dht.dart
@@ -319,7 +319,7 @@ Future<void> testWatchDHTValues(Stream<VeilidUpdate> updateStream) async {
     // as the watch's target
 
     final rcSet = await Veilid.instance.routingContext();
-    final rcWatch = await Veilid.instance.unsafeRoutingContext();
+    final rcWatch = await Veilid.instance.safeRoutingContext();
     try {
       // Make a DHT record
       var rec = await rcWatch.createDHTRecord(const DHTSchema.dflt(oCnt: 10));